Clean up + updated the genesis file content + fixed pool usage locations
