+++ indy-utils/src/main/kotlin/com/luxoft/blockchainlab/hyperledger/indy/IndyUser.kt
@@ -45,8 +45,6 @@ open class IndyUser {
     protected val wallet: Wallet
     protected val pool: Pool
 
-    constructor(wallet: Wallet, did: String? = null, didConfig: String = "{}") : this(TODO(), wallet, did, didConfig)
-
     constructor(pool: Pool, wallet: Wallet, did: String? = null, didConfig: String = "{}") {
 
         this.pool = pool
